feat/SW-38/refactor(Main): PostgreSQL 마이그레이션에 따른 Main(웹사이트 목록) 도메인 수정
* Note: 추후 확장 기능이기 때문에 일부만 마이그레이션

+++ src/main/java/com/web/SearchWeb/main/service/MainService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface MainService {
 
     //웹사이트 조회
-    Website selectWebsite(int websiteId);
+    Website selectWebsite(Long websiteId);
 
     //카테고리별 웹사이트 목록 조회
     List<Website> getListByCategory(String category);
